refactor(ddb): replace dynamoose w DDBSingleTable
Dynamoose has been replaced with custom DDB modeling tool
"DDBSingleTable", which wraps the DDB client and provides an improved
experience for managing a DDB table and its associated models. For more
information about the tool, please refer to its README, which is
currently located in src/lib/dynamoDB/DDBSingleTable.

The purpose of DDBSingleTable is to address the code-breaking problems
encountered when trying to use Dynamoose with just a single table which
ultimately made the tool's continued usage in this project a
fundamentally untenable proposition:
- Using 2+ Models with different "names" results in multiple tables.
- Using 2+ Models with the same "name" broke all Model `query` methods.

Aside from the code-breaking problems above, there were also a few
issues which were less impactful but worth mentioning:
- Limited typings
- Outdated/incorrect/incomplete documentation

A search for alternatives with a similar schema-based approach to db
modeling was not fruitful, hence the inception of DDBSingleTable.
In the future, DDBSingleTable will be moved to <root_dir>/packages,
which will require the repo to be configured using npm workspaces.

BREAKING_CHANGE: new modeling tool requires updates in src/models