build(codegen): add removal of leading tabs to node input str
Due to the way bash/TS handle leading whitespace in multiline strings, the TS output was improperly formatted (though the issue did not cause imports of the file content to fail). This fix causes the undesired leading whitespace to be stripped.

The `Date | null` UnionTypeNode is now also cached ahead of time to avoid instantiating a new one for every `Date | null` field.